refactor(components): isolate conditional tooltip logic - I233
